PE-72: Separated code into three disting files
* store handler functions in one file, positional functions in another and the command definitions in the third
* Ensures that no command will be run unless the players.json config exists
* moved all commands to lower case
* added command suggestion if passed command is wrong
* PE-72: fixed function naming error that was calling the wrong getfiles func
* indentation update
* updated documentation about broken file upload while pw is set
* fix: return in settime handler function
* fix: update logic in first settime format check
