+++ bin/handlerFunctions.js
@@ -1,95 +1,18 @@
 const fs = require('fs');
 const fsp = fs.promises;
-const FormData = require('form-data');
+const formData = require('form-data');
 let currentPath = require('path'); // for absolute path
 //const players = require('./players.json');
 const fetch = require('node-fetch');
 const os = require('os');
 const readline = require('readline');
 const fetchDigest = require('digest-fetch');
-const statusCodes = require('http-status');
-const crypto = require('crypto');
-const { argv } = require('process');
 const CONFIG_FILE_PATH = currentPath.join(os.homedir(), '.bsc', 'players.json');
-
-// Define error types
-const errorTypes = {
-    wrongPW: 0,
-    wrongUser: 1,
-    wrongIP: 2,
-    invalidURL: 3,
-    unknownError: 4,
-    badRequest: 5,
-    notFound: 6,
-    badPlayerName: 7,
-    unreachable: 8,
-}
-
-const miscErrorInfo = {
-    internalErrorMin: 500,
-    internalErrorMax: 599,
-    successMin: 200,
-    successMax: 299,
-    unreachableType: 'system',
-    unreachableNo: 'EHOSTUNREACH',
-}
-
-class ApiError extends Error {
-    constructor(message, status, contentType) {
-        super(message);
-        this.name = 'ApiError';
-        this.status = status;
-        this.contentType = contentType;
-    }
-}
-
-class playerNameError extends Error {
-    constructor(message, type) {
-        super(message);
-        this.name = 'playerNameError';
-        this.type = type;
-    }
-}
-
-function computeDigestHeader({
-  wwwAuthenticateHeader,
-  username,
-  password,
-  method,
-  url,
-}) {
-  const authParams = wwwAuthenticateHeader
-    .match(/(realm|qop|nonce|opaque)="([^"]+)"/gi)
-    .reduce((params, match) => {
-      const [_, key, value] = match.match(/([^=]+)="([^"]+)"/i);
-      params[key.toLowerCase()] = value;
-      return params;
-    }, {});
-
-  const nc = '00000001';
-  const cnonce = crypto.randomBytes(8).toString('hex');
-
-  const { realm, nonce, qop, opaque } = authParams;
-
-  const hash = (data) => crypto.createHash('md5').update(data).digest('hex');
-
-  const ha1 = hash(`${username}:${realm}:${password}`);
-  const ha2 = hash(`${method.toUpperCase()}:${url}`);
-  const response = hash(`${ha1}:${nonce}:${nc}:${cnonce}:${qop}:${ha2}`);
-
-  return `Digest username="${username}", realm="${realm}", nonce="${nonce}", uri="${url}", response="${response}", qop=${qop}, nc=${nc}, cnonce="${cnonce}", opaque="${opaque}"`;
-}
 
 // Handle commands
 // edit player
 function editPlayer(argv) {
   
-    // check if player exists
-    const players = JSON.parse(fs.readFileSync(CONFIG_FILE_PATH, 'utf8'));
-    if (typeof players[argv.playerName] === "undefined") {
-        errorHandler(new playerNameError('Player not found', errorTypes.badPlayerName));
-    }
-
     let playerName = argv.playerName;
     let playerSetUser = argv.username;
     let playerSetPass = argv.password;
@@ -115,21 +38,16 @@ function editPlayer(argv) {
             console.log('new username inputted');
             JSONdata[playerName].username = playerSetUser;
         }
-        if (argv.storage) {
-            console.log('new storage inputted');
-            JSONdata[playerName].storage = argv.storage;
-        }
 
         // stringify the new json object
         let newJSONdata = JSON.stringify(JSONdata, null, 2);
         // write to the file
         fs.writeFile(CONFIG_FILE_PATH, newJSONdata, 'utf8', (error) => {
-            if (error) {
-                console.error('Error writing file: ', error);
-                return;
-            } else if (argv.ipAddress || argv.password || argv.username || argv.storage) {
-                console.log('Player edited successfully');
-            }
+        if (error) {
+            console.error('Error writing file: ', error);
+            return;
+        }
+        console.log('Player edited successfully');
         });
     });
 }
@@ -148,14 +66,9 @@ function listPlayers() {
 // edit registry function
 async function editReg(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
     // create body
     let rawBody = JSON.stringify({ "value": argv.value });
 
@@ -171,48 +84,36 @@ async function editReg(argv) {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response);
     } catch (err) {
-        errorHandler(err);
+        console.log(err);
     }
 }
 
 // factory reset function
 async function factoryReset(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
     let requestOptions = {
         method: 'PUT',
         url: 'http://' + playerData[1] + '/api/v1/control/reboot',
-        headers: { 'Content-Type': 'application/json' },
-        body: rawBody
     }
+    requestOptions.body = { factory_reset: true };
 
     // send request
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response);
     } catch (err) {
-        errorHandler(err);
+        console.log(err);
     }
 }
 
 // set time function
 async function setTime(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
     let timezone = argv.timezone;
     let setDate = argv.date;
     let setTime = argv.time;
@@ -254,22 +155,16 @@ async function setTime(argv) {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log('Time set successfully: ' + response.data.result);
     }
-    catch (err) {
-        errorHandler(err);
+    catch (error) {
+        console.log(error);
     }
 }
 
 // get registry function
 async function getReg(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
     let section = argv.section;
     let key = argv.key;
     let requestOptions;
@@ -289,22 +184,16 @@ async function getReg(argv) {
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response.data.result.value);
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
 // set DWS function
 async function setDWS(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
     let onOff = argv.onOff;
     let rawBody;
 
@@ -342,22 +231,16 @@ async function setDWSsub(playerData, rawBody, onOff) {
         } else {
             console.log('set DWS failed');
         }
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
 // check DWS function
 async function checkDWS(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
 
     let requestOptions = {
         method: 'GET',
@@ -371,48 +254,36 @@ async function checkDWS(argv) {
         } else {
             console.log('DWS is disabled')
         }
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
 // get files func
 async function getFilesCom(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
     let playerPath = argv.path;
 
     let requestOptions = {
         method: 'GET',
-        url: 'http://' + playerData[1] + '/api/v1/files/' + playerData[3] + '/' + playerPath,
+        url: 'http://' + playerData[1] + '/api/v1/files/sd/' + playerPath,
     };
 
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response.data.result.files);
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
 // get time func
 async function getTime(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
 
     let requestOptions = {
         method: 'GET',
@@ -422,60 +293,52 @@ async function getTime(argv) {
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response.data.result);
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
 // delete file func
 async function deleteFile(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
     let playerPath = argv.file;
 
     let requestOptions = {
         method: 'DELETE',
-        url: 'http://' + playerData[1] + '/api/v1/files/' + playerData[3] + '/' + playerPath,
+        url: 'http://' + playerData[1] + '/api/v1/files/sd/' + playerPath,
     }
 
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(playerPath + ' deleted: ' + response.data.result.success);
     }
-    catch (err) {
-        errorHandler(err);
+    catch (error) {
+        console.log(error);
     }
 }
 
 // grab logs function
 async function getLogs(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let requestOptions = {
         method: 'GET',
-        url: 'http://' + playerData[1] + '/api/v1/logs',
+        url: 'http://' + playerIP + '/api/v1/logs',
     };
 
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response.data.result);
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
 }
 
@@ -495,7 +358,7 @@ async function handleRawRequest(argv) {
     };
 
     if (fileRaw != null) {
-        let form = new FormData();
+        let form = new formData();
         let fileToUpload = fs.createReadStream(fileRaw);
         console.log('Uploading file: ', fileRaw);
         form.append("file", fileToUpload, {filename: fileRaw});
@@ -504,8 +367,8 @@ async function handleRawRequest(argv) {
 
     try {
         let response = await requestFetch(requestOptions, 'admin', targetPasswordRaw);
-    } catch (err) {
-        errorHandler(err);
+    } catch (error) {
+        console.log(error);
     }
     if(rawResponseRaw) {
         console.log(response);
@@ -515,83 +378,92 @@ async function handleRawRequest(argv) {
 }
 
 // put file function
-async function pushFile(file, requestOptions, playerData) {
-  console.log(`pushing file: ${file}`);
-  const formData = new FormData();
-  formData.append('file', fs.createReadStream(file));
-  requestOptions.body = formData;
-  requestOptions.headers = formData.getHeaders();
-
-  try {
-    const response = await requestFetch(
-      requestOptions,
-      playerData[0],
-      playerData[2],
-      file
-    );
-    console.log(
-      `${response.data.result.results} uploaded: ${response.data.result.success}`
-    );
-  } catch (err) {
-    errorHandler(err);
-  }
-}
-
 async function push(argv) {
-  try {
-    const playerData = await pullData(argv); // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    const path = argv.FileDirectory;
-    const absPath = currentPath.resolve(path);
-    // check if file or directory
-    const isFile = await checkDir(path);
-    const requestOptions = {
-      method: "PUT",
-      url: `http://${playerData[1]}/api/v1/files/${
-        argv.location.length ? argv.location : "sd"
-      }`,
+    // get player data from argv
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
+
+    let requestOptions = {
+        method: 'PUT',
+        url: 'http://' + playerIP + '/api/v1/files/sd/' + argv.location,
     };
 
+    // check if file or directory
+    let path = argv.FileDirectory;
+    let absPath = currentPath.resolve(path);
+    let isFile;
+
+    isFile = await checkDir(path);
+    //console.log(isFile);
+
+    let files = [];
+    if(!isFile) {
+        console.log('getting files...');
+        try {
+        files = await getFiles(path);
+        }
+        catch (err) {
+        console.log('Error getting files from directory!');
+        console.log(err);
+        }
+    }
+
     if (isFile) {
-      await pushFile(absPath, requestOptions, playerData);
-      return;
+        // if file, push file
+        console.log('pushing file: ' + absPath);
+
+        let form = new formData();
+        let fileToUpload = fs.createReadStream(path);
+        form.append("file", fileToUpload, {filename: path});
+        requestOptions.body = form;
+
+        try {
+            let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
+            console.log(response.data.result.results + ' uploaded: ' + response.data.result.success);
+        } catch (err) {
+            console.log(err);
+        }
+    } else if (!isFile){
+        
+        // if directory, push directory
+        //console.log('pushing directory'); 
+        for (i = 0; i < files.length; i++) {
+
+            let fileToUpload = fs.createReadStream(files[i]);
+
+            let form = new formData();
+            form.append('file', fileToUpload, {filename: files[i]});
+            requestOptions.body = form;
+
+            console.log('Pushing ' + files[i]);
+
+            try {
+                let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
+                console.log(response.data.result.results + ' uploaded: ' + response.data.result.success);
+            } catch (err) {
+                console.log(err);
+            }
+        }
     }
-
-    console.log("getting files...");
-    const files = await getFiles(absPath);
-
-    for (const file of files) {
-      await pushFile(file, requestOptions, playerData);
-    }
-  } catch (err) {
-    console.log('An error occurred:', err);
-    errorHandler(err);
-  }
-}
-
-async function checkDir(path) {
-  try {
-    const stats = await fs.promises.stat(path);
-    return stats.isFile();
-  } catch (err) {
-    throw new Error('Failed to check if the path is a file or directory.');
-  }
 }
 
 // change password function
 async function changePW(argv) {
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let rawBody = JSON.stringify({
         "password": argv.newPassword,
-        "previous_password": playerData[2]
+        "previous_password": playerPW
     });
 
     let requestOptions = {
@@ -606,7 +478,7 @@ async function changePW(argv) {
         console.log('Password changed (player side): ' + response.data.result.success);
         //console.log(response);
     } catch (err) {
-        errorHandler(err);
+        console.log('Error changing password on player: ', err);
     }
 
     try {
@@ -621,7 +493,7 @@ async function changePW(argv) {
 
             // add new player
             JSONdata[argv.playerName] = {
-                ipAddress: playerData[1],
+                ipAddress: playerIP,
                 password: argv.newPassword
             }
 
@@ -644,18 +516,34 @@ async function changePW(argv) {
 async function checkPW(argv) {
 
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    // For next version:
+    // If password is set on player and not locally, this will return an error from requestFetch
+    // that should be handled within this function -> if certain error, then password is set on player
+    // so: 
+    /*
+        1. find out what error is returned when password is set on player and is not set, or set wrong, locally
+        a. call that error "authError", return as a boolean that is true if error is authError
+        2. if that error, then password is set on player and wrong/null locally
+        3. if that error AND password is set locally, then local password is wrong
+        a. if (authError && players[argv.playername].password !== '') -> password is wrong
+        4. if that error AND password not set locally -> must either turn player password off or set a password locally
+        a. if (authError && players[argv.playername].password == '') -> password is not set locally
+        5. if no error AND return object says password is blank (true), then password is not set on player
+        a. if (!authErro && response.data.result.password.isBlank) -> password is not set on player
+        6. if no error AND return object says password is not blank (false), then password is set on player and correct locally
+        a. if (!authError && !response.data.result.password.isBlank) -> password is set on player and correct locally
+    */
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let requestOptions = {
         method: 'GET',
-        url: 'http://' + playerData[1] + '/api/v1/control/dws-password',
+        url: 'http://' + playerIP + '/api/v1/control/dws-password',
     };
 
     try {
@@ -664,21 +552,8 @@ async function checkPW(argv) {
         //console.log(response.data.result.password);
         //console.log('Player has password set: ' + response.data.result.password.);
         console.log('Password is blank: ' + response.data.result.password.isBlank);
-        if (response.data.result.password.isBlank && (playerData[2] == '' || playerData[2] == undefined)) {
-            console.log('No password set locally, no password on player, you are good to go!')
-        } else if (!response.data.result.password.isBlank && (playerData[2] !== '' || playerData[2] !== undefined)) {
-            console.log('Password is set on the player, and your local password is set correctly, you are good to go!')
-        } else if (response.data.result.password.isBlank && (playerData[2] !== '' || playerData[2] !== undefined)) {
-            console.log('Password is not set on the player, but you have a password set locally. This will continue to work, but consider changing your password locally to be blank.')
-        }
-
     } catch (err) {
-        let errorType = checkpwErrorHandler(err);
-        if (errorType == errorTypes.wrongPW && (playerData[2] !== '' || playerData[2] !== undefined)) {
-            console.log('Your local password is wrong. Remember that the default password is the player serial number.');
-        } else if (errorType == errorTypes.wrongPW && (playerData[2] == '' || playerData[2] == undefined)) {
-            console.log('Your local password is not set, however the player has a password set');
-        }
+        console.log(err);
     }
 }
 
@@ -686,18 +561,16 @@ async function checkPW(argv) {
 async function reboot(argv) {
 
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let requestOptions = {
         method: 'PUT',
-        url: 'http://' + playerData[1] + '/api/v1/control/reboot',
+        url: 'http://' + playerIP + '/api/v1/control/reboot',
     };
 
     try {
@@ -705,7 +578,7 @@ async function reboot(argv) {
         //console.log(response);
         console.log('Player rebooted: ' + response.data.result.success);
     } catch (err) {
-        errorHandler(err);
+        console.log(err);
     }
 }
 
@@ -724,8 +597,7 @@ function addPlayer(argv) {
         JSONdata[argv.playerName] = {
             ipAddress: argv.ipAddress,
             password: argv.password,
-            username: argv.username,
-            storage: argv.storage
+            username: argv.username
         }
 
         // write new json object to file
@@ -768,25 +640,24 @@ function removePlayer(argv) {
 
 // get device info function
 async function getDeviceInfo(argv) {
+
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let requestOptions = {
         method: 'GET',
-        url: 'http://' + playerData[1] + '/api/v1/info',
+        url: 'http://' + playerIP + '/api/v1/info',
     };
     try {
         let response = await requestFetch(requestOptions, playerData[0], playerData[2]);
         console.log(response.data.result);
     } catch (err) {
-        errorHandler(err);
+        console.log(err);
     }
 }
 
@@ -794,18 +665,16 @@ async function getDeviceInfo(argv) {
 async function screenshot(argv) {
 
     // get player data from argv
-    let playerData;
-    try {
-        playerData = await pullData(argv);
-        // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
-    } catch (err) {
-        errorHandler(err);
-        return;
-    }
+    let playerData = await pullData(argv);
+    // playerData[0] = playerUser, [1] = playerIP, [2] = playerPW
+
+    let playerUser = playerData[0];
+    let playerIP = playerData[1];
+    let playerPW = playerData[2];
 
     let requestOptions = {
         method: 'POST',
-        url: 'http://' + playerData[1] + '/api/v1/snapshot',
+        url: 'http://' + playerIP + '/api/v1/snapshot',
     };
 
     try {
@@ -813,7 +682,7 @@ async function screenshot(argv) {
         //console.log(response);
         console.log('Screenshot taken! Location: ' + response.data.result.filename);
     } catch (err) {
-        errorHandler(err);
+        console.log(err);
     }
 }
 
@@ -836,94 +705,39 @@ function confirmDangerousCommand(prompt, callback) {
 }
 
 // generate players.json file if it doesn't exist
-async function generatePlayersJson() {
-    let fileExists = await fsp.access(CONFIG_FILE_PATH).then(() => true).catch(() => false);
-    try {
-        if (fileExists === true) {
-            //console.log('Players config file already exists');
-            return;
-        }
-
-        let playersDefault = {};
-        const rl = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout
-        });
-
-        const getPlayerName = () => {
-            return new Promise((resolve) => {
-                rl.question('Enter player name: ', resolve);
-            });
-        };
-        const getPlayerIP = () => {
-            return new Promise((resolve) => {
-                rl.question('Enter player IP address: ', resolve);
-            });
-        };
-        const getUsername = () => {
-            return new Promise((resolve) => {
-                rl.question('Enter player username: ', resolve);
-            });
-        };
-        const getPassword = () => {
-            return new Promise((resolve) => {
-                rl.question('Enter player password: ', resolve);
-            });
-        };
-        const getStorage = () => {
-            return new Promise((resolve) => {
-                rl.question('Enter player storage type: ', resolve);
-            });
-        };
-
-        await getPlayerName()
-            .then((playerName) => {
-                playersDefault[playerName] = {};
-                return getPlayerIP();
-            })
-            .then((ipAddress) => {
-                const playerName = Object.keys(playersDefault)[0];
-                playersDefault[playerName].ipAddress = ipAddress;
-                return getUsername();
-            })
-            .then((username) => {
-                const playerName = Object.keys(playersDefault)[0];
-                if (username === '') {
-                    playersDefault[playerName].username = 'admin';
-                } else {
-                    playersDefault[playerName].username = username;
-                }
-                return getPassword();
-            })
-            .then((password) => {
-                const playerName = Object.keys(playersDefault)[0];
-                playersDefault[playerName].password = password;
-                return getStorage();
-            })
-            .then((storage) => {
-                const playerName = Object.keys(playersDefault)[0];
-                if (storage === '') {
-                    playersDefault[playerName].storage = 'sd';
-                } else {
-                    playersDefault[playerName].storage = storage;
-                }
-                rl.close();
-                console.log(playersDefault);
-            })
-            .catch((error) => {
-                console.error(error);
-                rl.close();
-            });
-        const playersDefaultString = JSON.stringify(playersDefault, null, 2);
-
-        await fsp.mkdir(currentPath.join(os.homedir(), '.bsc'), { recursive: true });
-        await fsp.writeFile(CONFIG_FILE_PATH, playersDefaultString);
-
-        console.log('Players config file generated successfully');
-
-    } catch (err) {
-        console.error('Error generating players.json, ', err);
+function generatePlayersJson() {
+    if (fs.existsSync(CONFIG_FILE_PATH)) {
+        //console.log('Players config file already exists');
+        return;
     }
+
+    let playersDefault = {};
+    const rl = readline.createInterface({
+        input: process.stdin,
+        output: process.stdout
+    });
+
+    rl.question('Enter player name: ', (playerName) => {
+        playersDefault[playerName] = {};
+        rl.question('Enter player IP address: ', (ipAddress) => {
+        playersDefault[playerName].ipAddress = ipAddress;
+        rl.question('Enter player username: ', (username) => {
+            playersDefault[playerName].username = username;
+            rl.question('Enter player password: ', (password) => {
+            playersDefault[playerName].password = password;
+            rl.close();
+            let playersDefaultString = JSON.stringify(playersDefault, null, 2);
+
+            fs.mkdirSync(currentPath.join(os.homedir(), '.bsc'), { recursive: true });
+
+            fs.writeFileSync(CONFIG_FILE_PATH, playersDefaultString, (err) => {
+                if (err) throw err;
+                console.log('Players config file created');
+            });
+            });
+        });
+        });
+    });
 }
 
 // get player info from players.json
@@ -931,96 +745,63 @@ async function pullData(argv) {
 
     const players = JSON.parse(fs.readFileSync(CONFIG_FILE_PATH, 'utf8'));
 
-    if (typeof players[argv.playerName] === "undefined") {
-        throw new playerNameError('Player not found', errorTypes.badPlayerName);
-    } else {
-        let playerUser = players[argv.playerName].username;
-        let playerIP = players[argv.playerName].ipAddress;
-        let playerPW = players[argv.playerName].password;
-        let playerStorage = players[argv.playerName].storage;
+    let playerUser = players[argv.playerName].username;
+    let playerIP = players[argv.playerName].ipAddress;
+    let playerPW = players[argv.playerName].password;
 
-        let returnArr = [playerUser, playerIP, playerPW, playerStorage];
-        return returnArr;
-    }
+    let returnArr = [playerUser, playerIP, playerPW];
+    return returnArr;
 }
 
 // request fetch function
-async function requestFetch(requestOptions, user, pass, filePath) {
-  const { url, ...options } = requestOptions;
-  let succReturnContentType = 'application/json; charset=utf-8';
+async function requestFetch(requestOptions, user, pass) {
 
-  try {
-    let response = await fetch(url, options);
-    if (response.status === 401 && response.headers.has('www-authenticate')) {
-      const wwwAuthenticateHeader = response.headers.get('www-authenticate');
-      const digestHeader = await computeDigestHeader({
-        wwwAuthenticateHeader,
-        username: user,
-        password: pass,
-        method: options.method,
-        url,
-      });
-
-      let headers = {
-        ...options.headers,
-        Authorization: digestHeader,
-      };
-
-      let body = options.body;
-
-      if (filePath) {
-        const formData = new FormData();
-        formData.append('file', fs.createReadStream(filePath));
-        headers = {
-          ...formData.getHeaders(),
-          Authorization: digestHeader,
-        };
-        body = formData;
-      }
-
-      response = await fetch(url, {
-        method: options.method,
-        headers: headers,
-        body: body,
-      });
-    }
-
-    if (response.headers.get('content-type') === succReturnContentType) {
-      if (response.ok) {
-        return await response.json();
-      } else {
-        throw new ApiError(
-          'Response Error',
-          response.status,
-          response.headers.get('content-type')
-        );
-      }
+    if (pass !== "" && typeof pass !== "undefined") {
+        console.log('Password set, using digest auth')
+        //console.log(user);
+        if (typeof user === "undefined" || user === "") {
+            user = "admin";
+        }
+        let digestClient = new fetchDigest(user, pass);
+        try {  
+            let response = await digestClient.fetch(requestOptions.url, requestOptions);
+            let resData = await response.json();
+        return resData;
+        } catch (err) {
+            //console.error(err);
+            throw err;
+        }
     } else {
-      throw new ApiError(
-        'Unexpected content type in response',
-        response.status,
-        response.headers.get('content-type')
-      );
+        console.log('No password set, using no auth')
+        try {
+            let response = await fetch(requestOptions.url, requestOptions);
+            let resData = await response.json();
+            return resData;
+        } catch (err) {
+            //console.error(err);
+        throw err;
+        }
     }
-  } catch (err) {
-    throw err;
-  }
 }
 
 // check if location is directory function
 async function checkDir(path) {
-    try {
-        const stats = await fsp.stat(path);
-        if (stats.isFile()) {
-            return true;
-        } else if (stats.isDirectory()) {
-            return false;
-        } else {
-            throw new Error('Provided path is neither file nor directory');
+    return new Promise((resolve, reject) => {
+        fs.stat(path, (err, stats) => {
+        if (err) {
+            reject(err);
         }
-    } catch (err) {
-        throw err;
-    }
+        if (stats.isFile()) {
+            //console.log('file');
+            resolve(true);
+        } else if (stats.isDirectory()) {
+            //console.log('directory');
+            resolve(false);
+        } else {
+            reject(new Error('Provided path is neither file nor directory'));
+        }
+        });
+    });
 }
 
 // get files in directory function
@@ -1037,6 +818,7 @@ async function getFiles(path) {
         }
         return filesArr;
     } catch (err) {
+        console.error(err);
         throw err;
     }
 }
@@ -1051,84 +833,6 @@ function checkConfigExists() {
     } else {
         exists = true;
         return exists;
-    }
-}
-
-function errorHandler(err) {
-    //console.log(err);
-    /*
-    if (rawflag) {
-        console.error(err);
-    }
-    */
-    if (err.status == statusCodes.UNAUTHORIZED) {
-        //console.log(err)
-        console.log('\n');
-        //console.log('You have encountered an authorization error. This is most likely because your local password is incorrect, please check it and, if necessary, change it with the "editplayer" command. Example usage: bsc editplayer playerName -p playerPassword \n', err);
-        //console.log('For more info on your password, use the checkpw command')
-        console.log('Bad Auth: ', err);
-    } else if (err.status == statusCodes.BAD_REQUEST) {
-        //console.log(err)
-        console.log('\n');
-        //console.log('A bad request has  been sent. If you are not using the "raw" command, please open a github issue. If you are using "raw", please check your request options. \n', err);
-        console.log('Bad Request: ', err);
-    } else if (err.status == statusCodes.NOT_FOUND) {
-        //console.log(err)
-        console.log('\n');
-        //console.log('What you are looking for could not be found. If you are not using the "raw" command, please open a github issue. If you are using "raw", please check your request options. \n', err);
-        console.log('Not Found: ', err);
-    } else if (err.status >= miscErrorInfo.internalErrorMin && err.status <= miscErrorInfo.internalErrorMax) {
-        //console.log(err)
-        console.log('\n');
-        //console.log('You have encountered a server side error. If the problem persists, please open a github issue.', err);
-        console.log('Internal Server Error: ', err);
-    } else if (err.type == errorTypes.badPlayerName) {
-        console.log('Player name is invalid. Please check your player name and try again. \n', err);
-    } else if (err.type == statusCodes.unreachableType && err.errno == statusCodes.unreachableNo) {
-        //console.log('Your player is unreachable. Please check your locally configured IP and your/the player\'s internet access, and try again. \n', err);
-        console.log('Unreachable: ', err);
-    } else if (err.type = statusCodes.NOT_FOUND) {
-        //console.log('The resource you have attempted to interact with could not be found.\n', err);
-        console.log('Not Found: ', err);
-    }
-    else {
-        console.log('You have encountered an unknown error, please troubleshoot your issue (check player information locally, check if DWS is on, check if player is connected to internet, etc.)');
-        console.error(err);
-    }
-}
-
-// checkpw error handler
-function checkpwErrorHandler(err) {
-    if (err.status = statusCodes.UNAUTHORIZED) {
-        return errorTypes.wrongPW;
-    } else {
-        console.log('Unknown error, please try again');
-        return errorTypes.unknownError;
-    }
-}
-
-// check if help was called for specific commands, if so give extra info
-function helpChecker() {
-    // Get the arguments passed without actually parsing them
-    const rawArgs = process.argv.slice(2);
-    //console.log(rawArgs);
-    // check if help is invoked. Help can be invoked by specifically asking for it, or by 
-    // not providing any arguments
-    const isHelpInvoked = rawArgs.includes('--help') || rawArgs.includes('-h') || rawArgs.includes('help') || rawArgs.length == 1 || rawArgs.length == 2;
-
-    if (isHelpInvoked) {
-        if (rawArgs.includes('editplayer')) {
-            console.log('Example editplayer usage: ');
-            console.log('Set new password: bsc editplayer playerName -p playerPassword');
-            console.log('Set new username: bsc editplayer playerName -u playerUsername');
-            console.log('Set new IP and storage type: bsc editplayer playerName -i playerIP -s playerStorageType');
-        }
-        else if (rawArgs.includes('putfile')) {
-            console.log('Example putfile usage: ');
-            console.log('Push file: bsc putfile playerName <pathToFileLocally> [directoryOnPlayer]');
-            console.log('Push directory: bsc putfile playerName <pathToDirectoryLocally> [directoryOnPlayer]');
-            console.log('Please note that the \'locationOnPlayer\' argument is a DIRECTORY on the player, not the file name that you want to push to. If you do not specify a directory location, the file will be pushed to the root directory of the player storage.');
-        }
     }
 }
 
@@ -1156,6 +860,5 @@ module.exports = {
     getDeviceInfo,
     screenshot,
     generatePlayersJson,
-    checkConfigExists,
-    helpChecker
+    checkConfigExists
 };
